Refactor AdTree to cache table dimension sizes.
This commit introduces a new `dims` array in the `AdTree` class to store the number of categories for each variable in the table. This cache aims to improve performance by eliminating repetitive calculations for the number of categories. All code segments relying on `tableVariables.get(i).getNumCategories()` have been updated to use the new `dims` array.